fix: turn missing_mask into torch.float;

+++ pypots/data/base.py
@@ -205,7 +205,7 @@ class BaseDataset(Dataset):
         """
 
         X = self.X[idx].to(torch.float32)
-        missing_mask = ~torch.isnan(X)
+        missing_mask = (~torch.isnan(X)).to(torch.float32)
         X = torch.nan_to_num(X)
         sample = [
             torch.tensor(idx),
@@ -280,7 +280,7 @@ class BaseDataset(Dataset):
             self.file_handle = self._open_file_handle()
 
         X = torch.from_numpy(self.file_handle["X"][idx]).to(torch.float32)
-        missing_mask = ~torch.isnan(X)
+        missing_mask = (~torch.isnan(X)).to(torch.float32)
         X = torch.nan_to_num(X)
         sample = [
             torch.tensor(idx),
